update the folder structure
